+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changed
 
+- Removed greedy matching of dimension order in Bio-Formats Zarr output. Just check if OME-Zarr.
+
 ## 0.4.2
 
 ### Added